requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.31

# Data
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.4
pyarrow==15.0.2

# ML
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.4.0
catboost==1.2.5
imbalanced-learn==0.12.3

# Experiment tracking
mlflow==2.14.2

# Hyperparameter tuning
optuna==3.6.1

# Explainability
shap==0.45.1

# Visualization
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.2

# AI
google-genai>=1.0.0,<2.0.0

# Profiling
ydata-profiling==4.8.3
